% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{feldman_seurat}
\alias{feldman_seurat}
\title{A Seurat Object Sample}
\format{
An object of class \code{Seurat} with 78 rows and 16291 columns.
}
\usage{
feldman_seurat
}
\description{
Sade-Feldman dataset, analyzed and clustered using Seurat
}
\references{
Sade-Feldman et al.
}
\author{
Vy Nguyen \email{thi-tuong-vy.nguyen@meduniwien.ac.at}
}
\keyword{datasets}
