% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{jerby_seurat}
\alias{jerby_seurat}
\title{A Seurat Object Sample}
\format{
An object of class \code{Seurat} with 78 rows and 7186 columns.
}
\usage{
jerby_seurat
}
\description{
Sade-Feldman dataset, analyzed and clustered using Seurat
}
\references{
Jerby-Arnon et al.
}
\author{
Vy Nguyen \email{thi-tuong-vy.nguyen@meduniwien.ac.at}
}
\keyword{datasets}
